+++ packages/core/hardhat.config.js
@@ -21,13 +21,25 @@ const settings = {
   },
 };
 
+const settingsWithParisEVM = {
+  ...settings,
+  evmVersion: 'paris',
+};
+
+const proxyCompiler = {
+  version: require('./src/solidity-version.json'),
+  settings: settingsWithParisEVM,
+};
+
 function getNamespacedOverrides() {
   const contracts = fs.readdirSync(path.join(__dirname, 'contracts', 'test'));
   const namespacedContracts = contracts.filter(c => c.startsWith('Namespaced'));
   const overrides = {};
   for (const c of namespacedContracts) {
-    if (c !== 'NamespacedToModify07.sol') {
-      overrides[`contracts/test/${c}`] = { version: '0.8.20', settings: { ...settings, evmVersion: 'paris' } };
+    if (c === 'NamespacedToModify07.sol') {
+      overrides[`contracts/test/${c}`] = { version: '0.7.6', settings };
+    } else {
+      overrides[`contracts/test/${c}`] = { version: '0.8.20', settings: settingsWithParisEVM };
     }
   }
   return overrides;
@@ -49,6 +61,7 @@ module.exports = {
       { version: '0.7.6', settings },
       { version: '0.8.8', settings },
       { version: '0.8.9', settings },
+      proxyCompiler,
     ],
     overrides: getNamespacedOverrides(),
   },