+++ packages/plugin-hardhat/contracts/DeployContract.sol
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-pragma solidity ^0.8.9;
+pragma solidity ^0.8.20;
 
 import {Initializable as Initializable1} from "@openzeppelin/contracts/proxy/utils/Initializable.sol";
 import {Initializable as Initializable2} from "@openzeppelin/contracts-upgradeable/proxy/utils/Initializable.sol";