+++ packages/plugin-hardhat/contracts/Greeter50V3Proxiable.sol
@@ -1,6 +1,0 @@
-pragma solidity ^0.5.1;
-
-import "./GreeterV3.sol";
-import "./utils/Proxiable50.sol";
-
-contract Greeter50V3Proxiable is GreeterV3, Proxiable50 {}