+++ packages/plugin-hardhat/test/transparent-transfer-admin-ownership-happy-path.js
@@ -1,9 +1,10 @@
 const test = require('ava');
 
 const hre = require('hardhat');
-const { getManifestAdmin } = require('@openzeppelin/hardhat-upgrades/dist/admin.js');
 const { ethers, upgrades } = hre;
-const testAddress = '0x1E6876a6C2757de611c9F12B23211dBaBd1C9028';
+
+const TEST_ADDRESS = '0x1E6876a6C2757de611c9F12B23211dBaBd1C9028';
+const OWNABLE_ABI = ['function owner() view returns (address)'];
 
 test.before(async t => {
   t.context.Greeter = await ethers.getContractFactory('Greeter');
@@ -12,11 +13,13 @@ test.before(async t => {
 test('transferProxyAdminOwnership', async t => {
   // we need to deploy a proxy so we have a Proxy Admin
   const { Greeter } = t.context;
-  await upgrades.deployProxy(Greeter, ['Hello, Hardhat!'], { kind: 'transparent' });
+  const greeter = await upgrades.deployProxy(Greeter, ['Hello, Hardhat!'], { kind: 'transparent' });
 
-  const admin = await getManifestAdmin(hre);
-  await upgrades.admin.transferProxyAdminOwnership(testAddress);
+  await upgrades.admin.transferProxyAdminOwnership(await greeter.getAddress(), TEST_ADDRESS);
+
+  const adminAddress = await upgrades.erc1967.getAdminAddress(await greeter.getAddress());
+  const admin = await hre.ethers.getContractAt(OWNABLE_ABI, adminAddress);
   const newOwner = await admin.owner();
 
-  t.is(newOwner, testAddress);
+  t.is(newOwner, TEST_ADDRESS);
 });